+++ sandbox/libs/authn/src/main/java/org/opensearch/authn/realm/package-info.java
@@ -1,0 +1,7 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+/** Classes for the internal OpenSearch realm */
+package org.opensearch.authn.realm;