+++ server/src/main/java/org/opensearch/identity/noop/NoopSubject.java
@@ -8,9 +8,9 @@ package org.opensearch.identity.noop;
 import java.security.Principal;
 import java.util.Objects;
 
-import org.opensearch.identity.Subject;
-import org.opensearch.identity.AuthenticationToken;
-import org.opensearch.identity.Principals;
+import org.opensearch.authn.Subject;
+import org.opensearch.authn.AuthenticationToken;
+import org.opensearch.authn.Principals;
 
 /**
  * Implementation of subject that is always authenticated